+++ Messaging/Plugin.Messaging.WinStore/SmsTask.cs
@@ -11,10 +11,16 @@ namespace Plugin.Messaging
         #region ISmsTask Members
 
         public bool CanSendSms => false;
+        public bool CanSendSmsInBackground => false;
 
         public void SendSms(string recipient, string message)
         {
             throw new PlatformNotSupportedException("Sending SMS not supported on Windows Store");
+        }
+
+        public void SendSmsInBackground(string recipient, string message = null)
+        {
+            throw new PlatformNotSupportedException("Sending SMS in background not supported on Windows Store");
         }
 
         #endregion